Minor Windows compiler complaint fixes
http://codereview.appspot.com/5988072/



git-svn-id: http://skia.googlecode.com/svn/trunk@3644 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/gpu/GrContext.cpp
@@ -2108,7 +2108,8 @@ GrTexture* GrContext::gaussianBlur(GrTexture* srcTexture,
     SkRect srcRect(rect);
     scale_rect(&srcRect, 1.0f / scaleFactorX, 1.0f / scaleFactorY);
     srcRect.roundOut();
-    scale_rect(&srcRect, scaleFactorX, scaleFactorY);
+    scale_rect(&srcRect, static_cast<float>(scaleFactorX), 
+                         static_cast<float>(scaleFactorY));
     this->setClip(srcRect);
 
     const GrTextureDesc desc = {
